Fixed Check for status moves and renamed helper files
