+++ pokedex-react-flask-app/src/data/gmax_moves.json
@@ -25,7 +25,7 @@
     "pokemon": [{ "id": 10198, "name": "butterfree-gmax" }],
     "species_id": 12,
     "type": { "id": 7, "name": "bug" },
-    "effect": "Inflicts poison/, paralysis/, or sleep on all opponents"
+    "effect": "Inflicts poison, paralysis, or sleep on all opponents"
   },
   "10023": {
     "name": "g-max-volt-crash",
